pyproject.toml
[project]
name = "unrealzap"
version = "0.1.0"
description = "Unreal Tournament bug zapper"
authors = [{ name = "Danny Stewart", email = "danny@stewart.cc" }]
readme = "README.md"
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.2.4,<3.0.0)",
    "polykit (>=0.9.0,<0.10.0)",
    "pyalsaaudio (>=0.11.0,<0.12.0) ; sys_platform == 'linux'",
    "pygame (>=2.6.1,<3.0.0)",
    "scipy (>=1.15.2,<2.0.0)",
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.4"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
